convert-cli: rename controller to handler
